requirements.txt
asgiref==3.4.1
boto3==1.21.18
botocore==1.24.18
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
django-crispy-forms==1.14.0
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
numpy==1.22.2
pandas==1.4.1
Pillow==9.0.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
requests==2.26.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
